Corrected warning with C-style case

+++ GameEngine/Src/DynamicEntity.cpp
@@ -91,6 +91,6 @@ namespace Moo
 
 	void	DynamicEntity::evaporateHeatZone()
 	{
-			this->changeHealth(EVAPORATION_RATE / FPS_LIMIT);
+			this->changeHealth((float)(EVAPORATION_RATE / FPS_LIMIT));
 	}
 }
